finished assignment.

+++ todoapp/routers/todos.py
@@ -4,6 +4,7 @@ from database import engine, SessionLocal
 from typing import Annotated
 from sqlalchemy.orm import Session
 from pydantic import BaseModel, Field
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -17,6 +18,7 @@ def get_db():
         db.close()
         
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=1)
@@ -25,7 +27,7 @@ class TodoRequest(BaseModel):
     complete: bool = Field(default=False)
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency):
+async def read_all(user: user_dependency,db: db_dependency):
     """
     Read all todo records from the database.
     Parameters:
@@ -34,14 +36,19 @@ async def read_all(db: db_dependency):
     Here we are injecting the database session into the path operation function.
     This function will read all todo records from the database and return them.
     """
-    return db.query(Todos).all()
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+    return db.query(Todos).filter(Todos.owner_id == user["id"]).all()
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0), ):
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
     """
     This function will read a specific todo record by its ID from the database and return it.
     """
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first() #.first() returns the first result of the query and saves performance
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
 
     if todo_model is not None:
         return todo_model
@@ -49,22 +56,30 @@ async def read_todo(db: db_dependency, todo_id: int = Path(gt=0), ):
 
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
+async def create_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
     """
         This function will create a new todo record in the database.
     """
-    todo_model = Todos(**todo_request.model_dump())  # unpacking the fields from the Pydantic model to the SQLAlchemy model
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+
+    todo_model = Todos(**todo_request.model_dump(), owner_id=user["id"])  # unpacking the fields from the Pydantic model to the SQLAlchemy model
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, 
-                      todo_request: TodoRequest, 
-                      todo_id: int = Path(gt=0)):
+async def update_todo(
+        user: user_dependency,
+        db: db_dependency,
+        todo_request: TodoRequest,
+        todo_id: int = Path(gt=0)):
     """
     This function will update an existing todo record in the database.
     """
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
 
     if todo_model is None:
         raise HTTPException(status_code=404, detail=f"Todo with the id {todo_id} not found")
@@ -79,14 +94,17 @@ async def update_todo(db: db_dependency,
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
+async def delete_todo(user: user_dependency,db: db_dependency, todo_id: int = Path(gt=0)):
     """
     This function will delete a todo record from the database.
     """
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
 
     if todo_model is None:
         raise HTTPException(status_code=404, detail=f"Todo with the id {todo_id} not found")
-    
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).delete()
     db.commit()
